Ensure output CofM velocities have consistent units

+++ subhalo_properties.py
@@ -135,13 +135,13 @@ class SubhaloMasses(HaloProperty):
 
         # Compute centre of mass
         if total_mass_all > 0:
-            cofm_pos = (pos_times_mass / total_mass_all) % self.boxsize
+            cofm_pos = ((pos_times_mass / total_mass_all) % self.boxsize).astype(pos_dtype)
         else:
             cofm_pos = unyt.unyt_array((0,0,0), units=pos_units, dtype=pos_dtype)
 
         # Compute centre of mass velocity
         if total_mass_all > 0:
-            cofm_vel = vel_times_mass / total_mass_all
+            cofm_vel = (vel_times_mass / total_mass_all).astype(vel_dtype)
         else:
             cofm_vel = unyt.unyt_array((0,0,0), units=vel_units, dtype=vel_dtype)
